requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pandas==2.2.3
mysql-connector-python==8.0.26
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0 